combine filters into filters.py
